# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Travellers app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of travellers app"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get data about the current user",
        "operationId": "getMeUser",
        "description": "Receives data about the currently authorized user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved current user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found current user"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404-user"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by id",
        "operationId": "getUserById",
        "description": "Get a single user by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user",
            "example": "6881563901add19ee16fcff5"
          }
        ],
        "responses": {
          "200": {
            "description": "Get user by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found user with id 6881563901add19ee16fcff5"
                    },
                    "data": {
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404-user"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user by Id",
        "operationId": "updateUserData",
        "description": "Updates the user its id.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user to update",
            "example": "6912fb0ca18d66964cb28a9e"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Fields to update",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "New name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patch user by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched an user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404-user"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/{userId}/avatar": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update user avatar and add description",
        "operationId": "updateUserAvatar",
        "description": "Update the avatar for the currently authorized user and if description is required",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user to update",
            "example": "6912fb0ca18d66964cb28a9e"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "avatar"
                ],
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file for the new avatar"
                  },
                  "description": {
                    "type": "string",
                    "description": "Short text that describes the user",
                    "example": "Frontend developer and traveller"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile successfully updated!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Profile successfully updated!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "avatarUrl",
                        "id",
                        "description"
                      ],
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/user"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404-user"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "email",
          "password",
          "avatarUrl",
          "description",
          "onboardingCompleted",
          "savedStories"
        ],
        "properties": {
          "_id": {
            "description": "Users id",
            "type": "string",
            "example": "6881563901add19ee16fcff5"
          },
          "name": {
            "description": "User name",
            "type": "string",
            "example": "Дмитро Романенко"
          },
          "email": {
            "description": "User email",
            "type": "string",
            "example": "test123@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "qwerty"
          },
          "avatarUrl": {
            "description": "User avatar link",
            "type": "string",
            "example": "https://ftp.goit.study/img/harmoniq/users/6881563901add19ee16fcff5.webp"
          },
          "description": {
            "description": "Description about the user",
            "type": "string",
            "example": "User description"
          },
          "onboardingCompleted": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "savedStories": {
            "description": "IDs or populated list of user's saved stories",
            "type": "array",
            "items": {
              "description": "Stories Id",
              "type": "string",
              "example": "68498236a100312bea07902a"
            }
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "data": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Error object produced by http-errors (may contain name, message, details, etc.)",
                  "example": {
                    "name": "BadRequest",
                    "message": "Validation failed",
                    "details": [
                      "name is required",
                      "title must be a string"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized — missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden. The user does not have permission to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 403
                },
                "message": {
                  "type": "string",
                  "example": "You can only update your own avatar."
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "error"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal server erro"
                },
                "error": {
                  "type": "string",
                  "description": "Human-readable internal error message (from err.message)",
                  "example": "Database connection failed"
                }
              }
            }
          }
        }
      },
      "404-user": {
        "description": "User not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "example": {
                    "message": "User not found"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}